Notifications: Factor out error handling into a context manager
so we can reuse it later, and also avoid circular dependencies.
